+++ config/v3/loc_limits.yaml
@@ -267,6 +267,9 @@ code_limits:
       - path: "tests/vibe3/services/test_task_service_fresh_db.py"
         limit: 420
         reason: "TaskService fresh DB 测试集，覆盖 issue 链接、角色重分类、flow 降级、ref 选择等核心业务逻辑，包含完整的 mock setup 和参数化状态转换测试，拆分会破坏测试场景完整性"
+      - path: "tests/vibe3/commands/test_pr_show.py"
+        limit: 450
+        reason: "PR show 命令测试集，覆盖 PR 详情展示、inspect 生成、trace 输出、branch 推导逻辑等紧密耦合功能，拆分收益低"
 
 
   total_file_loc:
